requirements.txt
mysql-connector-python
yaml
loguru